+++ apps/fluent-tester/src/E2E/Avatar/pages/AvatarPageObject.ts
@@ -4,35 +4,37 @@ import {
   AVATAR_TEST_COMPONENT,
   AVATAR_SECONDARY_TEST_COMPONENT,
 } from '../../../TestComponents/Avatar/consts';
-import { BasePage, By } from '../../common/BasePage.win';
+import { BasePage, By } from '../../common/BasePage';
 
 export const enum AvatarComponentSelector {
   PrimaryComponent, //this._primaryComponent
   SecondaryComponent, //this._secondaryComponent
 }
 class AvatarPageObject extends BasePage {
-  getPrimaryComponentAttribute(attribute: string): string {
-    return this._primaryComponent.getAttribute(attribute);
-  }
-  getSecondaryComponentAttribute(attribute: string): string {
-    return this._secondaryComponent.getAttribute(attribute);
-  }
-  getAvatarAccessibilityLabel(componentSelector: AvatarComponentSelector): string {
-    return componentSelector == AvatarComponentSelector.SecondaryComponent
-      ? this._secondaryComponent.getAttribute('Name')
-      : this._primaryComponent.getAttribute('Name');
+  async getPrimaryComponentAttribute(attribute: string): Promise<string> {
+    return await this._primaryComponent.getAttribute(attribute);
   }
 
-  getAvatarAccessibilityHint(componentSelector: AvatarComponentSelector): string {
-    return componentSelector == AvatarComponentSelector.SecondaryComponent
-      ? this._secondaryComponent.getAttribute('HelpText')
-      : this._primaryComponent.getAttribute('HelpText');
+  async getSecondaryComponentAttribute(attribute: string): Promise<string> {
+    return await this._secondaryComponent.getAttribute(attribute);
   }
 
-  getAvatarAccessibilityRole(componentSelector: AvatarComponentSelector): string {
+  async getAvatarAccessibilityLabel(componentSelector: AvatarComponentSelector): Promise<string> {
     return componentSelector == AvatarComponentSelector.SecondaryComponent
-      ? this._secondaryComponent.getAttribute('ControlType')
-      : this._primaryComponent.getAttribute('ControlType');
+      ? await this._secondaryComponent.getAttribute('Name')
+      : await this._primaryComponent.getAttribute('Name');
+  }
+
+  async getAvatarAccessibilityHint(componentSelector: AvatarComponentSelector): Promise<string> {
+    return componentSelector == AvatarComponentSelector.SecondaryComponent
+      ? await this._secondaryComponent.getAttribute('HelpText')
+      : await this._primaryComponent.getAttribute('HelpText');
+  }
+
+  async getAvatarAccessibilityRole(componentSelector: AvatarComponentSelector): Promise<string> {
+    return componentSelector == AvatarComponentSelector.SecondaryComponent
+      ? await this._secondaryComponent.getAttribute('ControlType')
+      : await this._primaryComponent.getAttribute('ControlType');
   }
   /*****************************************/
   /**************** Getters ****************/