+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,36 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Wed, 31 Aug 2022 12:11:39 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Sep 2022 18:52:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.24.22
+
+Mon, 12 Sep 2022 18:52:11 GMT
+
+### Patches
+
+- Update shadow jest tests (78454019+lyzhan7@users.noreply.github.com)
+- Bump @fluentui-react-native/experimental-shadow to v0.2.16
+
+## 0.24.21
+
+Fri, 09 Sep 2022 23:54:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/experimental-shadow to v0.2.15
+
+## 0.24.20
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/icon to v0.13.6
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/pressable to v0.9.9
+- Bump @fluentui-react-native/text to v0.14.2
 
 ## 0.24.19