+++ packages/components/Menu/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @fluentui-react-native/menu
 
-This log was last generated on Wed, 31 Aug 2022 12:11:39 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Sep 2022 18:52:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.32
+
+Mon, 12 Sep 2022 18:52:11 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.22
+
+## 1.0.31
+
+Fri, 09 Sep 2022 23:54:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.21
+
+## 1.0.30
+
+Fri, 09 Sep 2022 01:08:39 GMT
+
+### Patches
+
+- Minor fixes for caching and adding new tokens (ruaraki@microsoft.com)
+
+## 1.0.29
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.21.9
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/text to v0.14.2
+- Bump @fluentui-react-native/button to v0.24.20
 
 ## 1.0.28