+++ packages/experimental/RadioGroup/CHANGELOG.md
@@ -1,8 +1,35 @@
 # Change Log - @fluentui-react-native/experimental-radio-group
 
-This log was last generated on Mon, 29 Aug 2022 23:21:44 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Sep 2022 21:48:46 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.0
+
+Mon, 12 Sep 2022 21:48:46 GMT
+
+### Minor changes
+
+- Added new Radio component with new framework. (gulnazsayed@microsoft.com)
+- Bump @fluentui-react-native/pressable to v0.9.9
+
+## 0.2.4
+
+Fri, 09 Sep 2022 21:53:29 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.10.10
+
+## 0.2.3
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.10.9
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/text to v0.14.2
 
 ## 0.2.2